requirements.txt
pdfplumber
pillow
sentence-transformers
faiss-cpu
flask
